endpoint per generazione documento
