+++ backend/template/template.service.ts
@@ -3,7 +3,6 @@ import { TemplateGenerator } from '../../services/TemplateGenerator';
 import { CreateTemplateDto, CreateTemplateResponse } from './dto/create-template.dto';
 import { UpdateContentDto, UpdateTemplateDto, UpdateTemplateResponse } from './dto/update-template.dto';
 import { DocumentConfig } from '../../interfaces/doc';
-import { GenerateDocumentDto, GenerateDocumentResponse } from './dto/generate-document.dto';
 import { DocumentGenerator } from '../../ContractGenerator';
 
 const DEFAULT_PATH_CONFIG = 'config.json';
@@ -72,25 +71,5 @@ export class TemplateService {
     const generator = new TemplateGenerator();
   }
 
-  async generateDocument(dto: GenerateDocumentDto): Promise<GenerateDocumentResponse> {
-    if (!dto.params) {
-      throw new BadRequestException('params is required to generate a document.');
-    }
 
-    const generator = new DocumentGenerator(DEFAULT_PATH_CONFIG);
-
-    try {
-      const document = await generator.generateDocument(dto.params);
-      if (!document) {
-        throw new Error('Document generation failed.');
-      }
-
-      return {
-        document,
-        filePath: dto.params.nomeFile,
-      };
-    } catch (error) {
-      throw new BadRequestException(`Unable to generate document: ${error instanceof Error ? error.message : error}`);
-    }
-  }
 }
